LayoutTests/platform/gtk/imported/w3c/web-platform-tests/pointerevents/persistentDeviceId/get-persistendeviceid-from-pointer-mouse-event.tentative-expected.txt

CONSOLE MESSAGE: Error: assert_equals: pointerover deviceId is 1 expected (number) 1 but got (undefined) undefined
CONSOLE MESSAGE: Error: assert_equals: pointerdown deviceId is 1 expected (number) 1 but got (undefined) undefined
CONSOLE MESSAGE: Error: assert_equals: pointermove deviceId is 1 expected (number) 1 but got (undefined) undefined
CONSOLE MESSAGE: Error: assert_equals: pointerup deviceId is 1 expected (number) 1 but got (undefined) undefined
CONSOLE MESSAGE: Error: assert_equals: click deviceId is 0 expected (number) 0 but got (undefined) undefined

Harness Error (FAIL), message = Error: assert_equals: click deviceId is 0 expected (number) 0 but got (undefined) undefined

PASS PointerEvent.persistentDeviceId
